factorize unit tests
